+++ apps/blog-app/src/main.server.ts
@@ -1,31 +1,8 @@
 import 'zone.js/node';
-import { enableProdMode } from '@angular/core';
-import { bootstrapApplication } from '@angular/platform-browser';
-import { renderApplication } from '@angular/platform-server';
-import { provideServerContext } from '@analogjs/router/server';
-import { ServerContext } from '@analogjs/router/tokens';
+import '@angular/platform-server/init';
+import { render } from '@analogjs/router/server';
 
 import { config } from './app/app.config.server';
 import { AppComponent } from './app/app.component';
 
-if (import.meta.env.PROD) {
-  enableProdMode();
-}
-
-export function bootstrap() {
-  return bootstrapApplication(AppComponent, config);
-}
-
-export default async function render(
-  url: string,
-  document: string,
-  serverContext: ServerContext
-) {
-  const html = await renderApplication(bootstrap, {
-    document,
-    url,
-    platformProviders: [provideServerContext(serverContext)],
-  });
-
-  return html;
-}
+export default render(AppComponent, config);